makaogift/src/main/java/kr/megaptera/makaogift/services/ProductService.java
package kr.megaptera.makaogift.services;

import kr.megaptera.makaogift.exceptions.ProductNotFound;
import kr.megaptera.makaogift.models.Product;
import kr.megaptera.makaogift.repositories.ProductRepository;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.List;

@Service
@Transactional
public class ProductService {
    private final ProductRepository productRepository;

    public ProductService(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    public List<Product> products() {
        List<Product> products = productRepository.findAll();
        return products;
    }

    public Product product(Long productId) {
        return productRepository.findById(productId)
                .orElseThrow(ProductNotFound::new);
    }
}
